Bundles are now stored in the ScreenRouter, and not as a static member of ScreenBundle.
This allows us to have more direct control over navigating through a sequential list of screens across multiple bundles. Also, it's the right place for it. Sorta. ScreenRouter is right now technically a "controller", but if it's holding the list of Bundles, in many ways, it's the Model as well.